* create db fail safe
* restructure tracker
