+++ forgebox/dbcore/__init__.py
@@ -7,6 +7,9 @@ Base = automap_base()
 try:
     engine = create_engine(SQLALCHEMY_DATABASE_URI)
 except:
+    import os
+    basedir = os.path.abspath(os.path.dirname(__file__))
+    DATADIR = os.path.expanduser("~/data")
     print("Create forge.db sqlite database file")
     from forge.app import db
     db.create_all()
